+++ plattis/rectangle.cc
@@ -7,8 +7,8 @@ SDL_Rect Rectangle::getDestRect(Camera* camera) const
 {
     return
     {
-        static_cast<int>(m_rect.x - camera->getX()),
-        static_cast<int>(m_rect.y - camera->getY()),
+        static_cast<int>(static_cast<float>(m_rect.x) - camera->getX()),
+        static_cast<int>(static_cast<float>(m_rect.y) - camera->getY()),
         m_rect.w,
         m_rect.h
     };
